Moved "ntersperse()" to it's own module.
