+++ src/rinoh/util/__init__.py
@@ -1,10 +1,26 @@
 from .etc import *
 from .rkd import RefKeyDictionary
 from .time import timed
+from .intersperse import intersperse
 
 
-__all__ = ['all_subclasses', 'intersperse', 'last', 'PeekIterator',
-           'posix_path', 'consumer', 'cached', 'cached_property',
-           'cached_generator', 'class_property', 'timed', 'Decorator',
-           'ReadAliasAttribute', 'NotImplementedAttribute', 'NamedDescriptor',
-           'WithNamedDescriptors', 'ContextManager', 'RefKeyDictionary']
+__all__ = [
+    "all_subclasses",
+    "intersperse",
+    "PeekIterator",
+    "posix_path",
+    "consumer",
+    "cached",
+    "cached_property",
+    "cached_generator",
+    "class_property",
+    "timed",
+    "Decorator",
+    "ReadAliasAttribute",
+    "NotImplementedAttribute",
+    "NamedDescriptor",
+    "WithNamedDescriptors",
+    "ContextManager",
+    "RefKeyDictionary",
+]
+
