+++ src/rinoh/util/etc.py
@@ -40,15 +40,6 @@ def all_subclasses(cls):
             yield subsubcls
 
 
-def intersperse(iterable, element):
-    """Generator yielding all elements of `iterable`, but with `element`
-    inserted between each two consecutive elements"""
-    iterable = iter(iterable)
-    yield next(iterable)
-    while True:
-        next_from_iterable = next(iterable)
-        yield element
-        yield next_from_iterable
 
 
 class PeekIterator(object):
